requirements.txt
pytest
pytest-playwright
playwright
pytest-xdist
allure-pytest
allure-python-commons
requests
urllib3
pytest-base-url
PyYAML